Check for and copy the contents of image/boot/<bootloader> in both classic and snap builds.
