+++ ubuntu_image/classic_builder.py
@@ -10,8 +10,6 @@ from tempfile import gettempdir
 from ubuntu_image.common_builder import AbstractImageBuilderState
 from ubuntu_image.helpers import (
      check_root_privilege, get_host_arch, live_build, run)
-from ubuntu_image.parser import (
-    BootLoader, FileSystemType, StructureRole)
 
 
 DEFAULT_FS = 'ext4'
@@ -131,60 +129,6 @@ class ClassicBuilder(AbstractImageBuilderState):
             userdata_file = os.path.join(cloud_dir, 'user-data')
             shutil.copy(self.cloud_init, userdata_file)
         super().populate_rootfs_contents()
-
-    def _populate_one_bootfs(self, name, volume):
-        for partnum, part in enumerate(volume.structures):
-            target_dir = os.path.join(volume.basedir, 'part{}'.format(partnum))
-            if part.role is StructureRole.system_boot:
-                volume.bootfs = target_dir
-                if (volume.bootloader is not BootLoader.uboot and
-                        volume.bootloader is not BootLoader.grub):
-                        raise ValueError(
-                            'Unsupported volume bootloader value: {}'.format(
-                                volume.bootloader))
-            if part.filesystem is FileSystemType.none:
-                continue
-            gadget_dir = os.path.join(self.unpackdir, 'gadget')
-            for content in part.content:
-                src = os.path.join(gadget_dir, content.source)
-                dst = os.path.join(target_dir, content.target)
-                if content.source.endswith('/'):
-                    # This is a directory copy specification.  The target
-                    # must also end in a slash.
-                    #
-                    # TODO: If this is a file instead of a directory, give
-                    # a useful error message instead of a traceback.
-                    #
-                    # TODO: We should assert this constraint in the parser.
-                    target, slash, tail = content.target.rpartition('/')
-                    if slash != '/' and tail != '':
-                        raise ValueError(
-                            'target must end in a slash: {}'.format(
-                                content.target))
-                    # The target of a recursive directory copy is the
-                    # target directory name, with or without a trailing
-                    # slash necessary at least to handle the case of
-                    # recursive copy into the root directory), so make
-                    # sure here that it exists.
-                    os.makedirs(dst, exist_ok=True)
-                    for filename in os.listdir(src):
-                        sub_src = os.path.join(src, filename)
-                        dst = os.path.join(target_dir, target, filename)
-                        if os.path.isdir(sub_src):
-                            shutil.copytree(sub_src, dst, symlinks=True,
-                                            ignore_dangling_symlinks=True)
-                        else:
-                            shutil.copy(sub_src, dst)
-                else:
-                    # TODO: If this is a directory instead of a file, give
-                    # a useful error message instead of a traceback.
-                    os.makedirs(os.path.dirname(dst), exist_ok=True)
-                    shutil.copy(src, dst)
-
-    def populate_bootfs_contents(self):
-        for name, volume in self.gadget.volumes.items():
-            self._populate_one_bootfs(name, volume)
-        super().populate_bootfs_contents()
 
     def generate_manifests(self):
         # After the images are built, we would also like to have some image
